requirements.txt
asn1crypto==1.5.1
boto3==1.40.49
botocore==1.40.49
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
filelock==3.20.0
idna==3.10
jmespath==1.0.1
numpy==2.2.6
packaging==25.0
pandas==2.3.3
platformdirs==4.5.0
pyarrow==21.0.0
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
snowflake-connector-python==4.0.0
sortedcontainers==2.4.0
tomlkit==0.13.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
